[new] - Stretch out suggestion API to allow dictionary, threshold, and limit params.
